# Add benchmark and performance fixes