--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,7 @@
 module github.com/ikawaha/dartsclone
 
 require (
+	github.com/euclidr/darts v0.0.0-20180401113647-e0859b23b68e
 	github.com/ikawaha/da v0.0.0-20141126165404-5556b9e515ca
 	github.com/schollz/progressbar/v2 v2.7.1
 )