--- a/internal/double_array_mmap_unix.go
+++ b/internal/double_array_mmap_unix.go
@@ -67,6 +67,17 @@ func openMmap(f *os.File, offset, length int) (*MmapedDoubleArray, error) {
 	return ret, nil
 }
 
+// Close deletes the mapped memory and closes the opened file.
+func (a *MmapedDoubleArray) Close() error {
+	if a.raw == nil {
+		return nil
+	}
+	data := a.raw
+	a.raw = nil
+	runtime.SetFinalizer(a, nil)
+	return syscall.Munmap(data)
+}
+
 func (a MmapedDoubleArray) at(i uint32) (unit, error) {
 	if _, err := a.r.Seek(int64(i*4), io.SeekStart); err != nil {
 		return 0, fmt.Errorf("seek error, %v", err)
@@ -80,26 +91,88 @@ func (a MmapedDoubleArray) at(i uint32) (unit, error) {
 
 // ExactMatchSearch searches TRIE by a given keyword and returns the id and it's length if found.
 func (a MmapedDoubleArray) ExactMatchSearch(key string) (id, size int, err error) {
-	return exactMatchSearch(a, key)
+	nodePos := uint32(0)
+	unit, err := a.at(nodePos)
+	if err != nil {
+		return -1, -1, err
+	}
+	for i := 0; i < len(key); i++ {
+		nodePos ^= unit.offset() ^ uint32(key[i])
+		unit, err = a.at(nodePos)
+		if err != nil {
+			return -1, -1, err
+		}
+		if unit.label() != key[i] {
+			return -1, 0, nil
+		}
+	}
+	if !unit.hasLeaf() {
+		return -1, 0, nil
+	}
+	unit, err = a.at(nodePos ^ unit.offset())
+	if err != nil {
+		return -1, -1, err
+	}
+	return int(unit.value()), len(key), nil
 }
 
 // CommonPrefixSearch finds keywords sharing common prefix in an input and returns the ids and it's lengths if found.
 func (a MmapedDoubleArray) CommonPrefixSearch(key string, offset int) (ids, sizes []int, err error) {
-	return commonPrefixSearch(a, key, offset)
+	nodePos := uint32(0)
+	unit, err := a.at(nodePos)
+	if err != nil {
+		return ids, sizes, err
+	}
+	nodePos ^= unit.offset()
+	for i := offset; i < len(key); i++ {
+		k := key[i]
+		nodePos ^= uint32(k)
+		unit, err := a.at(nodePos)
+		if err != nil {
+			return ids, sizes, err
+		}
+		if unit.label() != k {
+			break
+		}
+		nodePos ^= unit.offset()
+		if unit.hasLeaf() {
+			u, err := a.at(nodePos)
+			if err != nil {
+				return ids, sizes, err
+			}
+			ids = append(ids, int(u.value()))
+			sizes = append(sizes, i+1)
+		}
+	}
+	return ids, sizes, nil
 }
 
 // CommonPrefixSearchCallback finds keywords sharing common prefix in an input and callback with id and it's length.
 func (a MmapedDoubleArray) CommonPrefixSearchCallback(key string, offset int, callback func(id, size int)) error {
-	return commonPrefixSearchCallback(a, key, offset, callback)
-}
-
-// Close deletes the mapped memory and closes the opened file.
-func (a *MmapedDoubleArray) Close() error {
-	if a.raw == nil {
-		return nil
+	nodePos := uint32(0)
+	unit, err := a.at(nodePos)
+	if err != nil {
+		return err
 	}
-	data := a.raw
-	a.raw = nil
-	runtime.SetFinalizer(a, nil)
-	return syscall.Munmap(data)
+	nodePos ^= unit.offset()
+	for i := offset; i < len(key); i++ {
+		k := key[i]
+		nodePos ^= uint32(k)
+		unit, err := a.at(nodePos)
+		if err != nil {
+			return err
+		}
+		if unit.label() != k {
+			break
+		}
+		nodePos ^= unit.offset()
+		if unit.hasLeaf() {
+			u, err := a.at(nodePos)
+			if err != nil {
+				return err
+			}
+			callback(int(u.value()), i+1)
+		}
+	}
+	return nil
 }