--- a/trie_test.go
+++ b/trie_test.go
@@ -21,6 +21,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/euclidr/darts"
 	"github.com/ikawaha/da"
 )
 
@@ -37,6 +38,32 @@ func BenchmarkTRIE(b *testing.B) {
 	if err := scanner.Err(); err != nil {
 		b.Fatalf("unexpected scanner error, %v", err)
 	}
+	sort.Strings(keys)
+
+	b.Run("dartsclone", func(b *testing.B) {
+		trie, err := Open("./internal/_testdata/da_keys")
+		if err != nil {
+			b.Fatalf("unexpected error, dartsclone open, %v", err)
+		}
+		b.Run("exact match search", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				for _, v := range keys {
+					if id, _, err := trie.ExactMatchSearch(v); id < 0 || err != nil {
+						b.Fatalf("unexpected error, missing a keyword %v, id=%v, err=%v", v, id, err)
+					}
+				}
+			}
+		})
+		b.Run("common prefix match search", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				for _, v := range keys {
+					if ids, _, err := trie.CommonPrefixSearch(v, 0); len(ids) == 0 || err != nil {
+						b.Fatalf("unexpected error, missing a keyword %v, err=%v", v, err)
+					}
+				}
+			}
+		})
+	})
 
 	b.Run("github.com/ikawaha/da", func(b *testing.B) {
 		da := da.DoubleArray{}
@@ -63,16 +90,15 @@ func BenchmarkTRIE(b *testing.B) {
 			}
 		})
 	})
-	b.Run("dartsclone", func(b *testing.B) {
-		trie, err := Open("./internal/_testdata/da_keys")
-		if err != nil {
-			b.Fatalf("unexpected error, dartsclone open, %v", err)
-		}
+
+	b.Run("github.com/euclidr/darts", func(b *testing.B) {
+		builder := darts.DoubleArrayBuilder{}
+		builder.Build(keys)
 		b.Run("exact match search", func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				for _, v := range keys {
-					if id, _, err := trie.ExactMatchSearch(v); id < 0 || err != nil {
-						b.Fatalf("unexpected error, missing a keyword %v, id=%v, err=%v", v, id, err)
+					if _, ok := builder.ExactMatchSearch(v); !ok {
+						b.Fatalf("unexpected error, missing the keyword %v", v)
 					}
 				}
 			}
@@ -80,8 +106,8 @@ func BenchmarkTRIE(b *testing.B) {
 		b.Run("common prefix match search", func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				for _, v := range keys {
-					if ids, _, err := trie.CommonPrefixSearch(v, 0); len(ids) == 0 || err != nil {
-						b.Fatalf("unexpected error, missing a keyword %v, err=%v", v, err)
+					if ids := builder.CommonPrefixSearch(v); len(ids) == 0 {
+						b.Fatalf("unexpected error, missing a keyword %v", v)
 					}
 				}
 			}