… go.sum
… internal/_testdata/mmapbin_1_2_3_4_5
… internal/_testdata/mmapbin_20_1_2_3_4_5
… internal/double_array.go