before
```
goos: darwin
goarch: amd64
pkg: github.com/ikawaha/dartsclone
BenchmarkTRIEMmaped/exact_match_search-4         	       5	 293421881 ns/op	77895072 B/op	 8759265 allocs/op
BenchmarkTRIEMmaped/common_prefix_match_search-4 	       3	 421429137 ns/op	114470362 B/op	11507079 allocs/op
BenchmarkTRIE/dartsclone/exact_match_search-4    	      30	  53198414 ns/op	10427872 B/op	  325871 allocs/op
BenchmarkTRIE/dartsclone/common_prefix_match_search-4         	      10	 113095733 ns/op	38135028 B/op	 1965171 allocs/op
BenchmarkTRIE/github.com/ikawaha/da/exact_match_search-4      	      50	  40353694 ns/op	10427872 B/op	  325871 allocs/op
BenchmarkTRIE/github.com/ikawaha/da/common_prefix_match_search-4         	      20	  93832931 ns/op	38135032 B/op	 1965171 allocs/op
BenchmarkTRIE/github.com/euclidr/darts/exact_match_search-4              	     100	  21804084 ns/op	   65712 B/op	    1340 allocs/op
BenchmarkTRIE/github.com/euclidr/darts/common_prefix_match_search-4      	      30	  47886573 ns/op	13919292 B/op	  820990 allocs/op
PASS
```

after
```
goos: darwin
goarch: amd64
pkg: github.com/ikawaha/dartsclone
BenchmarkTRIEMmaped/exact_match_search-4         	      30	  33749317 ns/op	       0 B/op	       0 allocs/op
BenchmarkTRIEMmaped/common_prefix_match_search-4 	      20	  87990055 ns/op	27707160 B/op	 1639300 allocs/op
BenchmarkTRIE/dartsclone/exact_match_search-4    	      50	  25816305 ns/op	       0 B/op	       0 allocs/op
BenchmarkTRIE/dartsclone/common_prefix_match_search-4         	      20	  79275053 ns/op	27707157 B/op	 1639300 allocs/op
BenchmarkTRIE/github.com/ikawaha/da/exact_match_search-4      	      50	  39287358 ns/op	10427872 B/op	  325871 allocs/op
BenchmarkTRIE/github.com/ikawaha/da/common_prefix_match_search-4         	      10	 100244239 ns/op	38135027 B/op	 1965171 allocs/op
BenchmarkTRIE/github.com/euclidr/darts/exact_match_search-4              	     100	  21417760 ns/op	   65712 B/op	    1340 allocs/op
BenchmarkTRIE/github.com/euclidr/darts/common_prefix_match_search-4      	      30	  47706498 ns/op	13919291 B/op	  820990 allocs/op
```